requirements.txt
numpy
scipy
matplotlib
jplephem
pyephem
obspy
# you'll have to install this with:
# > pip install basemap --allow-external basemap --allow-unverified basemap
# or for Python 3:
# > pip3 install --user https://github.com/matplotlib/basemap/archive/master.zip
basemap
